remove ShiftScaleTransformerMulti in favor of TransformerMulti
