+++ src/opinf/pre/__init__.py
@@ -2,4 +2,5 @@
 """Tools for preprocessing snapshot data prior to compression."""
 
 from ._base import *
+from ._multi import *
 from ._shiftscale import *
