+++ tests/pre/test_shiftscale.py
@@ -408,7 +408,7 @@ class TestShiftScaleTransformer:
         assert np.all(Y == X)
 
         # Test centering.
-        st = self.Transformer(centering=True, scaling=None, byrow=False)
+        st = self.Transformer(centering=True, name="VaRiAbLe", verbose=True)
         Y = fit_transform_copy(st, X)
         assert isinstance(st.mean_, np.ndarray)
         assert st.mean_.shape == (X.shape[0],)
@@ -595,216 +595,3 @@ class TestShiftScaleTransformer:
         with pytest.raises(ValueError) as ex:
             st.inverse_transform(Y, locs=locs)
         assert ex.value.args[0] == "states_transformed not aligned with locs"
-
-
-class TestShiftScaleTransformerMulti:
-    """Test pre.ShiftScaleTransformerMulti."""
-
-    Transformer = opinf.pre.ShiftScaleTransformerMulti
-
-    def test_init(self):
-        """Test ShiftScaleTransformer.__init__()."""
-
-        # Centering.
-        stm = self.Transformer(2, centering=(True, False))
-        assert stm.centering == (True, False)
-        stm = self.Transformer(3, centering=True)
-        assert stm.centering == (True, True, True)
-        with pytest.raises(ValueError) as ex:
-            self.Transformer(3, centering=[True, False])
-        assert ex.value.args[0] == "len(centering) = 2 != 3 = num_variables"
-        with pytest.raises(TypeError) as ex:
-            self.Transformer(3, centering=100)
-
-        # Scaling.
-        stm = self.Transformer(2, scaling=("minmax", None))
-        assert stm.scaling == ("minmax", None)
-        stm = self.Transformer(2, scaling=[None, "maxabs"])
-        assert stm.scaling == (None, "maxabs")
-        stm = self.Transformer(3, scaling="standard")
-        assert stm.scaling == ("standard", "standard", "standard")
-        with pytest.raises(ValueError) as ex:
-            self.Transformer(3, scaling=(True, False))
-        assert ex.value.args[0] == "len(scaling) = 2 != 3 = num_variables"
-        with pytest.raises(TypeError) as ex:
-            self.Transformer(3, scaling=101)
-
-        # Verbose
-        stm.verbose = 1
-        assert stm.verbose is True
-        stm.verbose = 0
-        assert stm.verbose is False
-        assert all(st.verbose is False for st in stm.transformers)
-        stm.verbose = True
-        assert all(st.verbose is True for st in stm.transformers)
-
-    # Properties --------------------------------------------------------------
-    def test_mean(self, num_variables=4, varsize=7):
-        """Test ShiftScaleTransformerMulti.mean_."""
-        centerings = [False, True, False, True]
-        scalings = [None, None, "standard", "minmax"]
-        stm = self.Transformer(num_variables, centerings, scalings)
-        assert stm.mean_ is None
-
-        # Set the centering vectors.
-        stm.full_state_dimension = stm.num_variables * varsize
-        µs = [
-            np.random.randint(0, 100, stm.variable_size)
-            for _ in range(stm.num_variables)
-        ]
-        for i, µ in enumerate(µs):
-            if centerings[i]:
-                stm.transformers[i].mean_ = µ
-            if scalings[i]:
-                stm.transformers[i].scale_ = 0
-                stm.transformers[i].shift_ = 0
-            stm.transformers[i].n = stm.variable_size
-
-        # Validate concatenated mean_.
-        µµ = stm.mean_
-        assert isinstance(µµ, np.ndarray)
-        assert µµ.shape == (stm.full_state_dimension,)
-        for i, µ in enumerate(µs):
-            s = slice(i * stm.variable_size, (i + 1) * stm.variable_size)
-            if centerings[i]:
-                assert np.allclose(µµ[s], µ)
-            else:
-                assert np.allclose(µµ[s], 0)
-
-    # Magic methods -----------------------------------------------------------
-    def test_eq(self):
-        """Test ShiftScaleTransformerMulti.__eq__()."""
-        # Null transformers.
-        stm1 = self.Transformer(3)
-        assert stm1 != 100
-        stm2 = self.Transformer(2)
-        assert stm1 != stm2
-        stm2 = self.Transformer(3)
-        assert stm1 == stm2
-
-        # Mismatched attributes.
-        stm1 = self.Transformer(3, centering=True)
-        stm2 = self.Transformer(3, centering=False)
-        assert not (stm1 == stm2)
-        assert stm1 != stm2
-
-        stm1 = self.Transformer(3, scaling="minmax")
-        stm2 = self.Transformer(3, scaling="minmax")
-        assert stm1 == stm2
-        stm2 = self.Transformer(3, scaling=("minmax", "standard", "minmax"))
-        assert stm1 != stm2
-
-    # Main routines -----------------------------------------------------------
-    def test_mains(self, nx=50, k=400):
-        """Use ShiftScaleTransformerMulti.verify() to run tests."""
-
-        centerings = [
-            False,
-            True,
-            False,
-            False,
-            False,
-            False,
-            False,
-            True,
-            True,
-            True,
-            True,
-            True,
-        ]
-        scalings = [
-            None,
-            None,
-            "standard",
-            "minmax",
-            "minmaxsym",
-            "maxabs",
-            "maxabssym",
-            "standard",
-            "minmax",
-            "minmaxsym",
-            "maxabs",
-            "maxabssym",
-        ]
-        stm = self.Transformer(
-            len(centerings),
-            centering=centerings,
-            scaling=scalings,
-            verbose=True,
-        )
-
-        # Test dimension check.
-        with pytest.raises(ValueError) as ex:
-            stm.fit_transform(np.zeros(4 * len(centerings)))
-        assert ex.value.args[0] == "2D array required to fit transformer"
-
-        # Training checks.
-        Q = np.random.random((nx * stm.num_variables, k))
-        for method in (
-            stm.transform,
-            stm.transform_ddts,
-            stm.inverse_transform,
-        ):
-            with pytest.raises(AttributeError) as ex:
-                method(Q)
-            assert ex.value.args[0] == (
-                "transformer not trained (call fit() or fit_transform())"
-            )
-
-        # Verify the main routines.
-        Y = stm.fit_transform(Q, inplace=True)
-        assert Y is Q
-        t = np.linspace(0, 0.1, k)
-        stm.verify(Q, t)
-
-    # Persistence -------------------------------------------------------------
-    def test_save(self, nvar=15):
-        """Test ShiftScaleTransformerMulti.save()."""
-        # Clean up after old tests.
-        target = "_savetransformermultitest.h5"
-        if os.path.isfile(target):  # pragma: no cover
-            os.remove(target)
-
-        # Check file creation and on null transformation.
-        stm = self.Transformer(nvar)
-        stm.save(target)
-        assert os.path.isfile(target)
-
-        # Check non-null transformations.
-        scalings = {
-            None,
-            *TestShiftScaleTransformer.Transformer._VALID_SCALINGS,
-        }
-        for centering, scaling in itertools.product((True, False), scalings):
-            stm = self.Transformer(nvar, centering=centering, scaling=scaling)
-            X = np.random.randint(0, 100, (150, 17)).astype(float)
-            stm.fit_transform(X)
-            stm.save(target, overwrite=True)
-
-        # Clean up.
-        os.remove(target)
-
-    def test_load(self, nvar=5, nx=20, k=50):
-        """Test ShiftScaleTransformerMulti.load()."""
-        # Clean up after old tests.
-        target = "_loadtransformermultitest.h5"
-        if os.path.isfile(target):  # pragma: no cover
-            os.remove(target)
-
-        # Check that save() -> load() gives the same transformer.
-        stm = self.Transformer(nvar)
-        scalings = {
-            None,
-            *TestShiftScaleTransformer.Transformer._VALID_SCALINGS,
-        }
-        for centering, scaling in itertools.product((True, False), scalings):
-            stm = self.Transformer(nvar, centering=centering, scaling=scaling)
-            X = np.random.randint(0, 100, (150, 19)).astype(float)
-            stm.fit_transform(X)
-            stm.save(target, overwrite=True)
-            stm2 = self.Transformer.load(target)
-            assert len(stm2.transformers) == nvar
-            assert stm2 == stm
-
-        # Clean up.
-        os.remove(target)
